maint: initial commit adopting tentative governance model
