+++ .maint/CONTRIBUTORS.md
@@ -1,0 +1,13 @@
+# CONTRIBUTORS
+
+This document lists those who have made contributions to the Project.
+As per the contributor guidelines, they should be included by default in any publications derived from the Project.
+
+If you are new to the project, don't forget to add your name and affiliation to the list of contributors here! Our Welcome Bot will send an automated message reminding this to first-time contributors.
+Before every release, unlisted contributors will be invited again to add their names to this file (just in case they missed the automated message from our Welcome Bot).
+
+| **Lastname** | **Name** | **Handle** | **ORCID** | **Affiliation** |
+| --- | --- | --- | --- | --- |
+| Cieslak | Matthew | @mattcieslak | 0000-0002-1931-4734 | Perelman School of Medicine, University of Pennsylvania, PA, USA |
+| Kent | James D. | @jdkent | 0000-0002-4892-2659 | Neuroscience Program, University of Iowa |
+| Veraart | Jelle | @jelleveraart | 0000-0003-0781-0420 | School of Medicine, New York University |
