+++ .maint/FORMER.md
@@ -1,0 +1,9 @@
+# FORMER MEMBERS
+
+This document lists former contributors or maintainers who want to disengage from the Project, and seek to be dismissed in communications or future papers.
+By adding your name to this list you are giving up on all your responsibilities to the project.
+Should you desire to be considered back as a contributor or maintainer, please remove your name from this list and proceed as prescribed in the governance documents.
+
+| **Lastname** | **Name** | **Handle** | **ORCID** | **Affiliation** |
+| --- | --- | --- | --- | --- |
+|  | nrajamani3 | nrajamani3 |  |  |
