+++ .maint/MAINTAINERS.md
@@ -1,0 +1,18 @@
+# Maintainers
+
+This document lists the Maintainers of the project.
+Maintainers may be added once approved by the existing maintainers as described in the `../GOVERNANCE.md` document.
+By adding your name to this list you are agreeing to abide by the governance documents and to abide by all of the Organization's polices, including the code of conduct, trademark policy, and antitrust policy.
+If you are participating because of your affiliation with another organization (designated below), you represent that you have the authority to bind that organization to these policies.
+
+| **Lastname** | **Name** | **Handle** | **ORCID** | **Affiliation** |
+| --- | --- | --- | --- | --- |
+| Dickie | Erin W. | @edickie | 0000-0003-3028-9864 | Krembil Centre for Neuroinformatics, The Centre for Addiction and Mental Health, Dept. of Psychiatry, University of Toronto |
+| Esteban | Oscar | @oesteban | 0000-0001-8435-6191 | Dept. of Radiology, Lausanne University Hospital and University of Lausanne |
+| Joseph | Michael | @josephmje | 0000-0002-0068-230X | The Centre for Addiction and Mental Health |
+| Keshavan | Anisha | @akeshavan | 0000-0003-3554-043X | The University of Washington, eScience Institute |
+| Lerma-Usabiaga | Garikoitz | @garikoitz | 0000-0001-9800-4816 | Basque Center on Cognition, Brain and Language (BCBL)<br />Department of Psychology, Stanford University |
+| Mansour | Salim | @slimnsour | 0000-0002-1092-1650 | The Centre for Addiction and Mental Health |
+| Pisner | Derek | @dPys | 0000-0002-1228-0201 | Department of Psychology, University of Texas at Austin, TX, USA |
+| Richie-Halford | Adam | @richford | 0000-0001-9276-9084 | The University of Washington, eScience Institute |
+| Rokem | Ariel | @arokem | 0000-0003-0679-1985 | The University of Washington eScience Institute |
