+++ .maint/PIs.md
@@ -1,0 +1,15 @@
+# PRINCIPAL INVESTIGATORS
+
+This documents the key personnel who oversees the development of the Project and secures funding.
+The names in this file are designated by the Organization's TSC at the time of accepting the Project under the Organization.
+Changes to this file must be approved by the TSC.
+
+By having your name in this list you are agreeing to abide by the governance documents and to abide by all of the Organization's polices, including the code of conduct, trademark policy, and antitrust policy.
+If you are participating because of your affiliation with another organization (designated below), you represent that you have the authority to bind that organization to these policies.
+
+| **Lastname** | **Name** | **Handle** | **ORCID** | **Affiliation** | **Position** |
+| --- | --- | --- | --- | --- | --- |
+| Esteban | Oscar | @oesteban | 0000-0001-8435-6191 | Department of Radiology, Lausanne University Hospital and University of Lausanne | 2 |
+| Poldrack | Russell A. | @poldrack | 0000-0001-6755-0259 | Department of Psychology, Stanford University |  |
+| Rokem | Ariel | @arokem | 0000-0003-0679-1985 | The University of Washington eScience Institute, WA, USA | 1 |
+| Satterthwaite | Theodore D. |  | 0000-0001-7072-9399 | Perelman School of Medicine, University of Pennsylvania, PA, USA |  |
