+++ .maint/ROADMAP.md
@@ -1,0 +1,7 @@
+# ROADMAP
+
+This document states how the roadmap is built, discussed and monitored by the Maintainers, with the engagement of Contributors and PIs.
+
+For example, if the GitHub Projects feature is used, this document will contain a link to the corresponding Project and document who is responsible of managing the project, contacting Contributors and Maintainers to monitor the progress of activities, organizing meetings and discussions around the activities, etc.
+
+This document may also indicate how the *milestones* feature of the project is used and, as for Projects, who is responsible of what coordination actions, their frequence, etc.
