--- .maint/contributors.json
@@ -1,26 +1,0 @@
-[
-  {
-    "affiliation": "Perelman School of Medicine, University of Pennsylvania, PA, USA",
-    "name": "Cieslak, Matthew",
-    "orcid": "0000-0002-1931-4734",
-    "type": "Researcher"
-  },
-  {
-    "affiliation": "Neuroscience Program, University of Iowa",
-    "name": "Kent, James D.",
-    "orcid": "0000-0002-4892-2659",
-    "type": "Researcher"
-  },
-  {
-    "affiliation": "Department of Psychology, Stanford University",
-    "name": "Poldrack, Russell A.",
-    "orcid": "0000-0001-6755-0259",
-    "type": "Researcher"
-  },
-  {
-    "affiliation": "School of Medicine, New York University",
-    "name": "Veraart, Jelle",
-    "orcid": "0000-0003-0781-0420",
-    "type": "Researcher"
-  }
-]
