--- .maint/developers.json
@@ -1,50 +1,0 @@
-[
-  {
-     "affiliation": "Krembil Centre for Neuroinformatics, The Centre for Addiction and Mental Health, Dept. of Psychiatry, University of Toronto",
-     "name": "Dickie, Erin W.",
-     "orcid": "0000-0003-3028-9864"
-  },
-  {
-    "affiliation": "Dept. of Radiology, Lausanne University Hospital and University of Lausanne",
-    "name": "Esteban, Oscar",
-    "orcid": "0000-0001-8435-6191"
-  },
-  {
-    "affiliation": "The Centre for Addiction and Mental Health",
-    "name": "Joseph, Michael",
-    "orcid": "0000-0002-0068-230X"
-  },
-  {
-    "affiliation": "The University of Washington, eScience Institute",
-    "name": "Keshavan, Anisha",
-    "orcid": "0000-0003-3554-043X"
-  },
-  {
-    "affiliation": [
-      "Basque Center on Cognition, Brain and Language (BCBL)",
-      "Department of Psychology, Stanford University"
-    ],
-    "name": "Lerma-Usabiaga, Garikoitz",
-    "orcid": "0000-0001-9800-4816"
-  },
-  {
-     "affiliation": "The Centre for Addiction and Mental Health",
-     "name": "Mansour, Salim",
-     "orcid": "0000-0002-1092-1650"
-  },
-  {
-    "affiliation": "Department of Psychology, University of Texas at Austin, TX, USA",
-    "name": "Pisner, Derek",
-    "orcid": "0000-0002-1228-0201"
-  },
-  {
-    "affiliation": "The University of Washington, eScience Institute",
-    "name": "Richie-Halford, Adam",
-    "orcid": "0000-0001-9276-9084"
-  },
-  {
-    "affiliation": "The University of Washington eScience Institute",
-    "name": "Rokem, Ariel",
-    "orcid": "0000-0003-0679-1985"
-  }
-]
