--- .maint/former.json
@@ -1,5 +1,0 @@
-[
-  {
-    "name": "nrajamani3"
-  }
-]
