+++ .maint/requirements.txt
@@ -1,0 +1,2 @@
+click
+fuzzywuzzy
