+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,3 @@
+# Code of Conduct
+
+This Project belongs to the [*NiPreps* Community, and is under its code of conduct](https://www.nipreps.org/community/CODE_OF_CONDUCT/).
